requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.9
ClickUz==0.3
Django==4.0
django-payments==0.15.0
django-tinymce==3.4.0
djangorestframework==3.13.1
firebase-admin==5.2.0
gcloud==0.17.0
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
httplib2==0.20.2
idna==2.5
jws==0.1.3
msgpack==1.0.3
oauth2client==3.0.0
packaging==21.3
PaycomUz==1.2
Pillow==8.4.0
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
pyparsing==3.0.6
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2021.3
requests==2.18.0
requests-toolbelt==0.7.0
rsa==4.8
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.21.1
